Implement type casts and bitwise or operation.
Casts can be performed when popping and pushing data from the stack.
<x as u32> will pop a value from the stack, cast it to u32, and bind it in x.
[x as u32] will cast the value bound in x to a u32 and push it to the stack.

We can even cast to and from the str type by assuming the underlying data layout.
These are stored like a string view inside 8 bytes.
4 bytes for the index in a string store.
4 bytes for the length of the string.

The program
[0 as u64] . [32 as u64] . [5 as u64] . << . | . <x as str>
Will bind x as a string that starts at index 0 and has length 5.

These string casts are inherently unsafe, but can be very powerful.